Add filterNotNullAndDefined

+++ src/application/tenants/tenants.ts
@@ -1,10 +1,11 @@
-import {httpReqHandler} from 'application/errorHandling';
 import db from 'infrastructure/db';
+import {httpReqHandler} from 'application/errorHandling';
 import {BaseError} from 'application/errorHandling';
 import paymentService from 'infrastructure/paymentService';
-import {createTenant} from 'domain/entities';
 import authService from 'infrastructure/authService';
 import storageService from 'infrastructure/storageService';
+import {createTenant} from 'domain/entities';
+import {filterNotNullAndDefined} from 'utils/filterNotNullAndDefined';
 
 export const TenantsAdapter = () => {
   const create = httpReqHandler(async (req) => {
@@ -27,12 +28,9 @@ export const TenantsAdapter = () => {
 
   const retrieve = httpReqHandler(async (req) => {
     const {tenantId} = req.user;
-    let tenant = await db.tenants.retrieve(tenantId);
+    const tenant = await db.tenants.retrieve(tenantId);
     if (!tenant) throw new BaseError(404, 'Tenant Not Found');
-    if (tenant.theme) {
-      const url = await storageService.getUrl(tenant.theme);
-      tenant = {...tenant, theme: url};
-    }
+    if (tenant.theme) tenant.theme = await storageService.getUrl(tenant.theme);
 
     return {body: tenant};
   });
@@ -41,16 +39,16 @@ export const TenantsAdapter = () => {
     const {tenantId} = req.user;
     const tenant = await db.tenants.retrieve(tenantId);
     if (!tenant) throw new BaseError(404, 'Tenant Not Found');
-    if (tenant.stripeCustomerId) {
+    if (tenant.stripeCustomerId)
       await paymentService.customers.del(tenant.stripeCustomerId);
-    }
 
     deleteTenantFiles(tenantId);
 
     const users = await authService.listUsers(tenantId);
-    const promises = users.map(({email}) => authService.deleteUser(email));
+    if (!users.length) return {};
 
-    await Promise.all(promises || []);
+    const promises = users.map(({email}) => authService.deleteUser(email));
+    await Promise.all(promises);
     await db.tenants.del(tenantId);
 
     return {};
@@ -58,9 +56,9 @@ export const TenantsAdapter = () => {
 
   const deleteTenantFiles = async (tenantId: string) => {
     const files = await storageService.list(tenantId);
-    if (!files?.length) return;
+    if (!files?.length) return {};
 
-    const keys = files.map(({Key}) => Key || '');
+    const keys = files.map(({Key}) => Key).filter(filterNotNullAndDefined);
     await storageService.bulkDel(keys);
   };
